% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.ookla.broadband-package.R
\docType{package}
\name{dc.ookla.broadband-package}
\alias{dc.ookla.broadband}
\alias{dc.ookla.broadband-package}
\title{dc.ookla.broadband: Social Data Commons: Digital Communications: Ookla Broadband Speeds}
\description{
Allows user to easily get Ookla speed data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Zachary Alerte \email{rsu8zv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
